+++ remoulade/actor.py
@@ -16,7 +16,6 @@
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 import re
-import time
 
 from .broker import get_broker
 from .logging import get_logger