+++ remoulade/message.py
@@ -22,7 +22,7 @@ from collections import namedtuple
 from .broker import get_broker
 from .composition import pipeline
 from .encoder import Encoder, JSONEncoder
-from .results import Results
+from .results import FAILURE_RESULT, ErrorStored, Results
 
 #: The global encoder instance.
 global_encoder = JSONEncoder()
@@ -108,7 +108,7 @@ class Message(namedtuple("Message", (
         options.update(updated_options)
         return self._replace(**attributes, options=options)
 
-    def get_result(self, *, backend=None, block=False, timeout=None):
+    def get_result(self, *, backend=None, block=False, timeout=None, raise_on_error=True):
         """Get the result associated with this message from a result
         backend.
 
@@ -125,12 +125,15 @@ class Message(namedtuple("Message", (
             result.
           timeout(int): The maximum amount of time, in ms, to block
             while waiting for a result.
+          raise_on_error(bool): raise an error if the result stored in
+            an error
 
         Raises:
           RuntimeError: If there is no result backend on the default
             broker.
           ResultMissing: When block is False and the result isn't set.
           ResultTimeout: When waiting for a result times out.
+          ErrorStored: When the result is an error and raise_on_error is True
 
         Returns:
           object: The result.
@@ -144,7 +147,12 @@ class Message(namedtuple("Message", (
             else:
                 raise RuntimeError("The default broker doesn't have a results backend.")
 
-        return backend.get_result(self, block=block, timeout=timeout)
+        result = backend.get_result(self, block=block, timeout=timeout)
+        if result.error:
+            if raise_on_error:
+                raise ErrorStored(result.error)
+            return FAILURE_RESULT
+        return result.result
 
     def __str__(self):
         return "%s / %s" % (self.actor_name, self.message_id)