+++ remoulade/results/__init__.py
@@ -16,7 +16,9 @@
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 from .backend import Missing, ResultBackend
-from .errors import ResultError, ResultMissing, ResultTimeout
+from .constants import FAILURE_RESULT
+from .errors import ErrorStored, ResultError, ResultMissing, ResultTimeout
 from .middleware import Results
 
-__all__ = ["Missing", "ResultBackend", "ResultError", "ResultTimeout", "ResultMissing", "Results"]
+__all__ = ["ErrorStored", "Missing", "ResultBackend", "FAILURE_RESULT", "ResultError", "ResultTimeout",
+           "ResultMissing", "Results"]