+++ tests/middleware/test_prometheus.py
@@ -1,4 +1,5 @@
 import urllib.request as request
+from time import sleep
 
 import pytest
 
@@ -15,6 +16,9 @@ def test_prometheus_middleware_exposes_metrics():
         # And an instance of the prometheus middleware
         prom = prometheus.Prometheus()
         prom.after_process_boot(broker)
+
+        # If I wait for the server to start
+        sleep(0.01)
 
         # When I request metrics via HTTP
         with request.urlopen("http://127.0.0.1:9191") as resp: